improvement: unify prompt assembly
Co-locate all prompt composition behind llm/prompt/assembly.ts so history projection no longer owns the prompt seam. history-messages now delegates system-prompt rendering to renderSystemPrompt and session-context uses composeWithFragments instead of standalone PromptComposer.

- add assembly.ts (single seam for prompt fragments, layer rendering, and toolset prompt blocks)

- remove composer.ts (logic absorbed into assembly)

- wire history-messages, session-history, session-context and automations/generation through assembly
