+++ packages/server/src/llm/session-history.ts
@@ -5,8 +5,8 @@ import type { PrefixedString } from '@stitch/shared/id';
 import { getDb } from '@/db/client.js';
 import { messages } from '@/db/schema/sessions.js';
 import { buildHistoryMessages } from '@/llm/history-messages.js';
-import { getPromptUserContext } from '@/llm/prompt/builder.js';
-import type { PromptConfig } from '@/llm/prompt/builder.js';
+import { getPromptUserContext } from '@/llm/prompt/assembly.js';
+import type { PromptConfig } from '@/llm/prompt/assembly.js';
 import { readMemoryPromptContext } from '@/memory/snapshot.js';
 import { getSettings } from '@/settings/service.js';
 import { getSessionTodosPromptBlock } from '@/todos/service.js';
