+++ packages/server/src/llm/session-summary.ts
@@ -14,19 +14,17 @@ import * as Log from '@/lib/log.js';
 import { addCacheControlToMessages } from '@/llm/cache-control.js';
 import { CompactionModelNotFoundError } from '@/llm/errors.js';
 import { buildHistoryMessages } from '@/llm/history-messages.js';
-import { getPromptUserContext } from '@/llm/prompt/builder.js';
+import { getPromptUserContext } from '@/llm/prompt/assembly.js';
 import { getProviderOptions } from '@/llm/provider-options.js';
 import { createProvider } from '@/llm/provider/provider.js';
 import { resolveCheapModel } from '@/llm/resolve-cheap-model.js';
 import { mapAIError, toStreamErrorDetails } from '@/llm/stream/ai-error-mapper.js';
-import { getSessionToolsetState } from '@/llm/stream/session-toolsets.js';
 import * as LocalModels from '@/models/llm/local.js';
 import * as Models from '@/models/llm/registry.js';
 import type { LlmProviderCredentials } from '@/provider/config/schema.js';
 import { getSettings } from '@/settings/service.js';
 import { getSessionTodosPromptBlock } from '@/todos/service.js';
 import { findBrowserProtectOverrides } from '@/tools/toolsets/browser/prune-policy.js';
-import { getToolset } from '@/tools/toolsets/registry.js';
 import { recordLlmUsage } from '@/usage/ledger.js';
 import { estimate } from '@/utils/token.js';
 import { normalizeUsage } from '@/utils/usage.js';
@@ -427,17 +425,6 @@ export async function compact(input: {
   } finally {
     activeCompactions.delete(sessionId);
   }
-}
-
-export function buildActiveToolsetInstructionsBlock(sessionId: PrefixedString<'ses'>): string {
-  const activeIds = getSessionToolsetState(sessionId).active.map((entry) => entry.id);
-  const instructionBlocks = activeIds
-    .map((id) => getToolset(id))
-    .filter((ts): ts is NonNullable<ReturnType<typeof getToolset>> => !!ts?.instructions)
-    .map((ts) => `### ${ts.name} Toolset Instructions\n${ts.instructions}`)
-    .join('\n\n');
-
-  return instructionBlocks ? `\n\n## Active Toolset Instructions\n\n${instructionBlocks}` : '';
 }
 
 /**
